dog-viewer - move profile into ProfileSection
